frontend/requirements.txt
flask==2.0.1
werkzeug==2.0.3
pyyaml==6.0.1
python-dotenv==1.0.0
jinja2==3.0.3
